Updates readme.  Adds trained model as well as the one-hot encoding for that model.  Renames some classes for improved clarity.
